+++ auth.py
@@ -13,27 +13,22 @@
 #  limitations under the License.
 
 import functools
-import ldap
 
+import ldap
 from flask import (
     Blueprint, flash, g, redirect, render_template, request, session, url_for, current_app
 )
-from werkzeug.security import check_password_hash, generate_password_hash
-
-from .db import find_one, insert_one
-
-from bson.objectid import ObjectId
-from .config import Config
-
-from oic.oic import Client
-from oic.utils.authn.client import CLIENT_AUTHN_METHOD
-from oic.oic.message import AuthorizationResponse
-from oic.oic.message import RegistrationResponse
 from oic import rndstr
+from oic.oic import Client
+from oic.oic.message import RegistrationResponse, AuthorizationResponse, ClaimsRequest, Claims
+from oic.utils.authn.client import CLIENT_AUTHN_METHOD
 from oic.utils.http_util import Redirect
+from werkzeug.security import check_password_hash
 
-bp = Blueprint('auth', __name__, url_prefix=Config.URL_PREFIX+'/auth')
-# current_app.config.from_pyfile('config.py', silent=True)
+from .config import Config
+from .db import find_one
+
+bp = Blueprint('auth', __name__, url_prefix=Config.URL_PREFIX + '/auth')
 # Create OIDC client
 client = Client(client_authn_method=CLIENT_AUTHN_METHOD)
 # Get authentication provider details by hitting the issuer URL.
@@ -43,6 +38,7 @@ info = {"client_id": Config.CLIENT_ID, "client_secret": Config.CLIENT_SECRET, "r
 client_reg = RegistrationResponse(**info)
 client.store_registration_info(client_reg)
 
+
 def check_login(view):
     @functools.wraps(view)
     def wrapped_view(**kwargs):
@@ -51,7 +47,9 @@ def check_login(view):
             if Config.ROLE.get(access) is not None:
                 return redirect(Config.ROLE.get(access)[1])
         return view(**kwargs)
+
     return wrapped_view
+
 
 def role_required(role):
     def decorator(view):
@@ -63,11 +61,13 @@ def role_required(role):
             else:
                 if Config.ROLE.get(access) is not None:
                     if Config.ROLE.get(access)[0] <= Config.ROLE.get(role)[0] and access != role:
-                       return redirect(Config.ROLE.get(access)[1])
+                        return redirect(Config.ROLE.get(access)[1])
                 else:
                     return redirect(url_for("auth.login"))
                 return view(**kwargs)
+
         return decorated_function
+
     return decorator
 
 
@@ -86,6 +86,7 @@ def login_db(username, password, error):
 
     flash(error)
     return False
+
 
 def login_ldap(username, password, error):
     ldap_hostname = current_app.config['LDAP_HOSTNAME']
@@ -128,14 +129,20 @@ def login_ldap(username, password, error):
     flash(error)
     return False
 
+
 def login_shi():
-    # session["mode"] = "shibboleth"
     session["state"] = rndstr()
     session["nonce"] = rndstr()
+    claims_request = ClaimsRequest(
+        userinfo=Claims(
+            uiucedu_uin={"essential": True}
+        )
+    )
     args = {
         "client_id": client.client_id,
         "response_type": "code",
         "scope": Config.SCOPES,
+        "claims": claims_request,
         "nonce": session["nonce"],
         "redirect_uri": client.registration_response["redirect_uris"][0],
         "state": session["state"]
@@ -144,61 +151,13 @@ def login_shi():
     login_url = auth_req.request(client.authorization_endpoint)
     return Redirect(login_url)
 
-    # if request.method == 'POST':
-    #     username = request.form['username']
-    #     password = request.form['password']
-    #     error = None
-    #
-    #     if current_app.config['LDAP_ON']: # log in using LDAP
-    #         result = login_ldap(username, password, error)
-    #     else:
-    #         result = login_db(username, password, error)
-    #
-    #     if result: # log in successful
-    #         if 'source-login' in request.form:
-    #             session['mode'] = 'source'
-    #             return redirect(url_for('event.source', sourceId=0))
-    #         if 'user-login' in request.form:
-    #             session['mode'] = 'user'
-    #             return redirect(url_for('user_events.user_events'))
-    #
-    #     flash(error)
-    #
-    # if session.get('mode') == 'source':
-    #     return redirect(url_for('event.source', sourceId=0))
-    # if session.get('mode') == 'user':
-    #     return redirect(url_for('user_events.user_events'))
-    # return render_template('auth/login.html')
-
-# @bp.route('/register', methods=('GET', 'POST'))
-# def register():
-#     if request.method == 'POST':
-#         username = request.form['username']
-#         password = request.form['password']
-#         error = None
-#
-#         if not username:
-#             error = 'Username is required.'
-#         elif not password:
-#             error = 'Password is required.'
-#         elif find_one('user', condition={"username": username}):
-#             error = 'User {} is already registered.'.format(username)
-#
-#         if error is None:
-#             password_hash = generate_password_hash(password)
-#             insert_one('user', document={"username": username, "password_hash": password_hash})
-#             return redirect(url_for('auth.login'))
-#
-#         flash(error)
-#
-#     return render_template('auth/register.html')
-
 
 @bp.route('/login')
 @check_login
 def login():
     if Config.LOGIN_MODE == "shibboleth":
         return login_shi()
+
 
 @bp.route('/callback')
 def callback():
@@ -215,39 +174,39 @@ def callback():
     token_response = client.do_access_token_request(state=authentication_response["state"],
                                                     request_args=args,
                                                     authn_method="client_secret_basic")
-    user_info = client.do_user_info_request(state=authentication_response["state"])
+    user_info = client.do_user_info_request(state=authentication_response["state"]).to_dict()
 
-    if "uiucedu_is_member_of" not in user_info.to_dict():
+    if "uiucedu_is_member_of" not in user_info:
         session.clear()
         return redirect(url_for("home.home", error="You don't have permission to login the event manager"))
-    rokwireAuth = list(filter(
-        lambda x: "urn:mace:uiuc.edu:urbana:authman:app-rokwire-service-policy-" in x, 
-        user_info.to_dict()["uiucedu_is_member_of"]
+    rokwire_auth = list(filter(
+        lambda x: "urn:mace:uiuc.edu:urbana:authman:app-rokwire-service-policy-" in x,
+        user_info["uiucedu_is_member_of"]
     ))
-    if len(rokwireAuth) == 0:
+    if len(rokwire_auth) == 0:
         return redirect(url_for("auth.login"))
     else:
         # fill in user information
-        session["name"] = user_info.to_dict()["name"]
-        session["email"] = user_info.to_dict()["email"]
+        session["name"] = user_info["name"]
+        session["email"] = user_info["email"]
         # check for corresponding privilege
-        isUserAdmin = False 
-        isSourceAdmin = False
-        for tag in rokwireAuth:
+        is_user_admin = False
+        is_source_admin = False
+        for tag in rokwire_auth:
             if "rokwire em user events admins" in tag:
-                isUserAdmin = True
+                is_user_admin = True
             if "rokwire em calendar events admins" in tag:
-                isSourceAdmin = True
+                is_source_admin = True
         # TODO: we are storing cookie by our own but not by code, may change it later
-        if isUserAdmin and isSourceAdmin:
+        if is_user_admin and is_source_admin:
             session["access"] = "both"
             session.permanent = True
             return redirect(url_for("auth.select_events"))
-        elif isUserAdmin:
+        elif is_user_admin:
             session["access"] = "user"
             session.permanent = True
             return redirect(url_for("user_events.user_events"))
-        elif isSourceAdmin:
+        elif is_source_admin:
             session["access"] = "source"
             session.permanent = True
             return redirect(url_for("event.source", sourceId=0))
@@ -255,11 +214,6 @@ def callback():
             session.clear()
             return redirect(url_for("home.home", error="You don't have permission to login the event manager"))
 
-    # if "member" in user_info.to_dict()["eduperson_affiliation"]:
-    #     return redirect(url_for('user_events.user_events'))
-    # else:
-    #     return redirect(url_for('event.source', sourceId=0))
-    # return user_info.to_json()
 
 @bp.route('/select-events', methods=['GET', 'POST'])
 @role_required("both")
@@ -275,27 +229,21 @@ def select_events():
             return render_template("auth/select-events.html", no_search=True)
     return render_template("auth/select-events.html", no_search=True)
 
+
 @bp.before_app_request
 def load_logged_in_user_info():
     if session.get("access") is None:
         g.user = None
     else:
-        g.user = {}
-        g.user["access"] = session["access"]
-        g.user["username"] = session["name"]
+        g.user = {"access": session["access"], "username": session["name"]}
 
-    # user_id = ObjectId(session.get('user_id'))
-
-    # if user_id is None:
-    #     g.user = None
-    # else:
-    #     g.user = find_one('user', condition={'_id': user_id})
 
 @bp.route('/logout')
 @role_required('either')
 def logout():
     session.clear()
     return redirect(url_for('home.home'))
+
 
 def login_required(view):
     @functools.wraps(view)